lambda/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
boto3==1.40.16
botocore==1.40.16
certifi==2025.8.3
charset-normalizer==3.4.3
distro==1.9.0
feedparser==6.0.11
google-generativeai==0.8.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
jmespath==1.0.1
pydantic==2.11.7
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
requests==2.32.5
s3transfer==0.13.1
sgmllib3k==1.0.0
six==1.17.0
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
